Refactor the code in ModuleOperationUtil
