+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/common/ImportsAcceptor.java
@@ -20,6 +20,7 @@ import io.ballerina.compiler.syntax.tree.ImportDeclarationNode;
 import io.ballerina.projects.Project;
 import org.ballerinalang.langserver.codeaction.CodeActionModuleId;
 import org.ballerinalang.langserver.common.utils.CommonUtil;
+import org.ballerinalang.langserver.common.utils.ModuleOperationUtil;
 import org.ballerinalang.langserver.commons.DocumentServiceContext;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.Range;
@@ -83,7 +84,7 @@ public class ImportsAcceptor {
                 String pkgName;
 
                 String moduleName = codeActionModuleId.moduleName();
-                String modulePrefix = CommonUtil.escapeModuleName(moduleName).replaceAll(".*\\.", "");
+                String modulePrefix = ModuleOperationUtil.escapeModuleName(moduleName).replaceAll(".*\\.", "");
                 if (!codeActionModuleId.modulePrefix().isEmpty() &&
                         !modulePrefix.equals(codeActionModuleId.modulePrefix())) {
                     moduleName = codeActionModuleId.moduleName() + " as " + codeActionModuleId.modulePrefix();
