+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/completions/providers/context/ImportDeclarationNodeContext.java
@@ -31,6 +31,7 @@ import io.ballerina.projects.ProjectKind;
 import org.ballerinalang.annotation.JavaSPIService;
 import org.ballerinalang.langserver.LSPackageLoader;
 import org.ballerinalang.langserver.common.utils.CommonUtil;
+import org.ballerinalang.langserver.common.utils.ModuleOperationUtil;
 import org.ballerinalang.langserver.commons.BallerinaCompletionContext;
 import org.ballerinalang.langserver.commons.LanguageServerContext;
 import org.ballerinalang.langserver.commons.completion.LSCompletionItem;
@@ -192,7 +193,7 @@ public class ImportDeclarationNodeContext extends AbstractCompletionProvider<Imp
             String orgName = pkg.packageOrg().value();
             String pkgName = pkg.packageName().value();
             if (orgName.equals(Names.BALLERINA_INTERNAL_ORG.getValue())
-                    || CommonUtil.matchingImportedModule(ctx, pkg).isPresent()) {
+                    || ModuleOperationUtil.matchingImportedModule(ctx, pkg).isPresent()) {
                 // Avoid suggesting the ballerinai org name
                 return;
             }
@@ -238,7 +239,7 @@ public class ImportDeclarationNodeContext extends AbstractCompletionProvider<Imp
                     .collect(Collectors.joining("."));
             String qualifiedModuleName = module.moduleName().packageName().value()
                     + Names.DOT + moduleNamePart;
-            if (CommonUtil.matchingImportedModule(ctx, "", qualifiedModuleName).isPresent()) {
+            if (ModuleOperationUtil.matchingImportedModule(ctx, "", qualifiedModuleName).isPresent()) {
                 continue;
             }
             String label = currentPackageName + "." + moduleNamePart;
@@ -270,7 +271,7 @@ public class ImportDeclarationNodeContext extends AbstractCompletionProvider<Imp
                     + module.moduleName().moduleNamePart();
             if (module.isDefaultModule()
                     || (currentModule.isPresent() && module.moduleId().equals(currentModule.get().moduleId()))
-                    || CommonUtil.matchingImportedModule(context, "", qualifiedModuleName).isPresent()) {
+                    || ModuleOperationUtil.matchingImportedModule(context, "", qualifiedModuleName).isPresent()) {
                 return;
             }
 
@@ -298,7 +299,7 @@ public class ImportDeclarationNodeContext extends AbstractCompletionProvider<Imp
             String packageName = ballerinaPackage.packageName().value();
             String insertText;
             if (orgName.equals(ballerinaPackage.packageOrg().value()) && !addedPkgNames.contains(packageName)
-                    && CommonUtil.matchingImportedModule(context, ballerinaPackage).isEmpty()) {
+                    && ModuleOperationUtil.matchingImportedModule(context, ballerinaPackage).isEmpty()) {
                 if (orgName.equals(Names.BALLERINA_ORG.value)
                         && packageName.startsWith(Names.LANG.getValue() + Names.DOT.getValue())) {
                     insertText = getLangLibModuleNameInsertText(packageName);
